refactor: rename `models` folder to `entities`
