+++ src/app/entities/Book.ts
