+++ src/app/entities/Store.ts
