+++ src/app/entities/User.ts
